feat: Attempt to rename the service to products
