--- README.template.md
@@ -1,5 +1,0 @@
-# ${placeholder-for-name}
-
-The current project was created using the template over at [template-go](https://github.com/Knoblauchpilze/template-go).
-
-This README was auto-generated and needs to be updated with an actual description of what this project does.
