+++ cmd/products/.gitignore
