+++ cmd/products/configs/.gitignore
