+++ cmd/products/configs/config-template-dev.yml
