+++ cmd/products/internal/config.go
