+++ cmd/products/internal/config_test.go
